requirements.txt
openai==0.28
python-dotenv
pinecone
langchain_openai
langchain
langchain-community
lark